requirements.txt
flask
crewai
langchain
openai
python-dotenv